+++ All_lessons/lesson17/GISFeature.cs
@@ -1,0 +1,55 @@
+using System;
+using System.Collections;
+//using System.Collections;
+using System.Collections.Generic;
+using System.Data;
+using System.Data.OleDb;
+using System.Deployment.Application;
+using System.Diagnostics.Eventing.Reader;
+using System.Drawing;
+using System.IO;
+using System.Linq.Expressions;
+using System.Runtime.InteropServices;
+using System.Runtime.InteropServices.WindowsRuntime;
+using System.Security.Cryptography.X509Certificates;
+using System.Security.Policy;
+using System.Text;
+using System.Windows.Forms;
+
+namespace My.GIS
+{
+/*class GISLine
+    {
+        List<GISVertex> AllVertexs;
+    }
+    class GISPolygon
+    {
+        List<GISVertex> AllVertexs;
+    }*/
+    public class GISFeature
+    {
+        public int ID;
+        public GISSpatial spatialPart;
+        public GISAttribute attributePart;
+        public bool Selected = false;
+        public GISFeature(GISSpatial spatial, GISAttribute attribute)
+        {
+            spatialPart = spatial;
+            attributePart = attribute;
+        }
+        public void Draw(Graphics graphics, MapAndClientConverter view, bool drawAttributeOrNot, int index,GISThematic thematic)
+        {
+            spatialPart.Draw(graphics, view, Selected,thematic);
+            if (drawAttributeOrNot)
+            {
+                attributePart.Draw(graphics, view, spatialPart.centroid, index);
+            }
+        }
+        public Object GetAttribute(int index)
+        {
+            return attributePart.GetValue(index);
+        }
+
+
+    }
+}
