+++ All_lessons/lesson17/GISPoint.cs
@@ -1,0 +1,68 @@
+using System;
+using System.Collections;
+//using System.Collections;
+using System.Collections.Generic;
+using System.Data;
+using System.Data.OleDb;
+using System.Deployment.Application;
+using System.Diagnostics.Eventing.Reader;
+using System.Drawing;
+using System.IO;
+using System.Linq.Expressions;
+using System.Runtime.InteropServices;
+using System.Runtime.InteropServices.WindowsRuntime;
+using System.Security.Cryptography.X509Certificates;
+using System.Security.Policy;
+using System.Text;
+using System.Windows.Forms;
+
+namespace My.GIS
+{
+/*
+     class GISPoint
+    {
+        public GISVertex location;
+        public string attribute;
+        //constructor
+        public GISPoint(GISVertex gISVertex, string pointAttribute)
+        {
+            location = gISVertex;
+            attribute = pointAttribute;
+        }
+        public void DrawPoint(Graphics graphics)
+        {
+            graphics.FillEllipse(new SolidBrush(Color.Red),
+                new Rectangle((int)(location.x)-3,(int)(location.y)-3,6,6));
+        }
+        public void DrawAttribute(Graphics graphics)
+        {
+            graphics.DrawString(attribute, new Font("LXGW", 20), new SolidBrush(Color.Green),
+                new PointF((int)location.x, (int)(location.y)));
+        }
+        public double VertexToPoint(GISVertex anotherVertex)
+        {
+            return location.DistanceToVertex(anotherVertex);
+        }
+
+    }
+    */
+    public class GISPoint : GISSpatial
+    {
+        public GISPoint(GISVertex vertext)
+        {
+            centroid = vertext;
+            mapExtent = new GISMapExtent(vertext, vertext);
+        }
+        public override void Draw(Graphics graphics, MapAndClientConverter view, bool Selected, GISThematic thematic)
+        {
+            Point screenPoint = view.ToScreenPoint(centroid);
+
+            graphics.FillEllipse(new SolidBrush(Selected ? GISConst.SelectedPointColor : thematic.InsideColor), new Rectangle((int)screenPoint.X - thematic.Size, (int)screenPoint.Y - thematic.Size, thematic.Size * 2, thematic.Size*2));
+            graphics.DrawEllipse(new Pen(new SolidBrush(thematic.OutsideColor)), new Rectangle(screenPoint.X - thematic.Size, screenPoint.Y - thematic.Size, thematic.Size * 2, thematic.Size * 2));;
+        }
+        public double GetDistanceThisPointToVertex(GISVertex vertex)
+        {
+            return centroid.GetDistanceThisVToV(vertex);
+        }
+    }
+}
