+++ All_lessons/lesson17/lesson17.csproj
@@ -67,17 +67,26 @@
     <Compile Include="Form1.Designer.cs">
       <DependentUpon>Form1.cs</DependentUpon>
     </Compile>
+    <Compile Include="GISAttribute.cs" />
+    <Compile Include="GISFeature.cs" />
+    <Compile Include="GISLine.cs" />
+    <Compile Include="GISMapExtent.cs" />
     <Compile Include="GISPanel.cs">
       <SubType>UserControl</SubType>
     </Compile>
     <Compile Include="GISPanel.Designer.cs">
       <DependentUpon>GISPanel.cs</DependentUpon>
     </Compile>
+    <Compile Include="GISPoint.cs" />
+    <Compile Include="GISPolygon.cs" />
+    <Compile Include="GISSelect.cs" />
     <Compile Include="GISTools.cs" />
+    <Compile Include="GISVertex.cs" />
     <Compile Include="Layer.cs" />
     <Compile Include="MyFiles.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ShapefileTools.cs" />
     <Compile Include="Thematic.cs" />
     <EmbeddedResource Include="AttributeWindow.resx">
       <DependentUpon>AttributeWindow.cs</DependentUpon>
